requirements.txt
python-dotenv
requests
flet